+++ docs/components/plugins/api/APISectionProps.tsx
@@ -1,9 +1,11 @@
+import { css } from '@emotion/react';
 import React from 'react';
 
 import { InlineCode } from '~/components/base/code';
+import { H4 } from '~/components/base/headings';
 import { LI, UL } from '~/components/base/list';
 import { P } from '~/components/base/paragraph';
-import { H2, H4, H4Code } from '~/components/plugins/Headings';
+import { H2, H3, H3Code } from '~/components/plugins/Headings';
 import {
   CommentTagData,
   DefaultPropsDefinitionData,
@@ -24,6 +26,10 @@ export type APISectionPropsProps = {
 };
 
 const UNKNOWN_VALUE = '...';
+
+const PROP_LIST_ELEMENT_STYLE = css`
+  padding: 0;
+`;
 
 const extractDefaultPropValue = (
   { comment, name }: PropData,
@@ -50,14 +56,17 @@ const renderInheritedProp = (ip: TypeDefinitionData) => {
   );
 };
 
-const renderInheritedProps = (data: TypeDefinitionData[] | undefined): JSX.Element | undefined => {
+const renderInheritedProps = (
+  data: TypeDefinitionData[] | undefined,
+  exposeInSidebar?: boolean
+): JSX.Element | undefined => {
   const inheritedProps = data?.filter((ip: TypeDefinitionData) => ip.type === 'reference') ?? [];
   if (inheritedProps.length) {
     return (
-      <div>
-        <H4>Inherited Props</H4>
+      <>
+        {exposeInSidebar ? <H3>Inherited Props</H3> : <H4>Inherited Props</H4>}
         <UL>{inheritedProps.map(renderInheritedProp)}</UL>
-      </div>
+      </>
     );
   }
   return undefined;
@@ -65,7 +74,8 @@ const renderInheritedProps = (data: TypeDefinitionData[] | undefined): JSX.Eleme
 
 const renderProps = (
   { name, type }: PropsDefinitionData,
-  defaultValues?: DefaultPropsDefinitionData
+  defaultValues?: DefaultPropsDefinitionData,
+  exposeInSidebar?: boolean
 ): JSX.Element => {
   const baseTypes = type.types
     ? type.types?.filter((t: TypeDefinitionData) => t.declaration)
@@ -79,17 +89,23 @@ const renderProps = (
     <div key={`props-definition-${name}`}>
       <UL>
         {propsDeclarations?.map(prop =>
-          prop ? renderProp(prop, extractDefaultPropValue(prop, defaultValues)) : null
+          prop
+            ? renderProp(prop, extractDefaultPropValue(prop, defaultValues), exposeInSidebar)
+            : null
         )}
       </UL>
-      {renderInheritedProps(type.types)}
+      {renderInheritedProps(type.types, exposeInSidebar)}
     </div>
   );
 };
 
-const renderProp = ({ comment, name, type, flags }: PropData, defaultValue?: string) => (
-  <LI key={`prop-entry-${name}`}>
-    <H4>{name}</H4>
+const renderProp = (
+  { comment, name, type, flags }: PropData,
+  defaultValue?: string,
+  exposeInSidebar?: boolean
+) => (
+  <LI key={`prop-entry-${name}`} customCss={exposeInSidebar ? PROP_LIST_ELEMENT_STYLE : undefined}>
+    {exposeInSidebar ? <H3>{name}</H3> : <H4>{name}</H4>}
     <P>
       {flags?.isOptional && <span css={STYLES_SECONDARY}>Optional&emsp;&bull;&emsp;</span>}
       <span css={STYLES_SECONDARY}>Type:</span> <InlineCode>{resolveTypeName(type)}</InlineCode>
@@ -115,14 +131,14 @@ const APISectionProps: React.FC<APISectionPropsProps> = ({
         <H2 key="props-header">{header}</H2>
       ) : (
         <>
-          <H4Code key={`${header}-props-header`}>
+          <H3Code key={`${header}-props-header`}>
             <InlineCode>{header}</InlineCode>
-          </H4Code>
+          </H3Code>
           <br />
         </>
       )}
       {data.map((propsDefinition: PropsDefinitionData) =>
-        renderProps(propsDefinition, defaultProps)
+        renderProps(propsDefinition, defaultProps, header === 'Props')
       )}
     </>
   ) : null;