+++ packages/expo-asset/src/LocalAssets.web.ts
@@ -1,4 +1,0 @@
-export function getLocalAssetUri(hash: string, type: string | null): string | null {
-  // noop on web
-  return null;
-}