+++ packages/expo-facebook/src/Facebook.ts
@@ -14,6 +14,7 @@ export {
   FacebookLoginResult,
   FacebookOptions,
   FacebookAuthenticationCredential,
+  FacebookInitializationOptions,
   PermissionResponse,
   PermissionStatus,
   PermissionExpiration,
@@ -46,6 +47,13 @@ const androidPermissionsResponse: PermissionResponse = {
   status: PermissionStatus.GRANTED,
 };
 
+/**
+ * Asks for permissions to use data for tracking the user or the device.
+ *
+ * > iOS: it requires the NSUserTrackingUsageDescription message added to the info.plist.
+ *
+ * @returns A promise that resolves to an object of type [PermissionResponse](#permissionresponse).
+ */
 export async function requestPermissionsAsync(): Promise<PermissionResponse> {
   if (Platform.OS === 'android') {
     return Promise.resolve(androidPermissionsResponse);
@@ -57,6 +65,13 @@ export async function requestPermissionsAsync(): Promise<PermissionResponse> {
   return await ExponentFacebook.requestPermissionsAsync();
 }
 
+/**
+ * Checks application's permissions for using data for tracking the user or the device.
+ *
+ * >iOS: it requires the NSUserTrackingUsageDescription message added to the info.plist.
+ *
+ * @returns A promise that resolves to an object of type [PermissionResponse](#permissionresponse).
+ */
 export async function getPermissionsAsync(): Promise<PermissionResponse> {
   if (Platform.OS === 'android') {
     return Promise.resolve(androidPermissionsResponse);
@@ -68,6 +83,13 @@ export async function getPermissionsAsync(): Promise<PermissionResponse> {
   return await ExponentFacebook.getPermissionsAsync();
 }
 
+/**
+ * Prompts the user to log into Facebook and grants your app permission to access their Facebook data. On iOS and Android, if the Facebook app isn't installed then a web view will be used to authenticate.
+ *
+ * @param options A map of options:
+ *  - **permissions (array)** -- An array specifying the permissions to ask for from Facebook for this login. The permissions are strings as specified in the [Facebook API documentation](https://developers.facebook.com/docs/permissions/reference). The default permissions are ['public_profile', 'email'].
+ * @returns If the user or Facebook cancelled the login, returns { type: 'cancel' }. Otherwise, returns { type: 'success' } & [FacebookAuthenticationCredential](#facebookauthenticationcredential).
+ */
 export async function logInWithReadPermissionsAsync(
   options: FacebookOptions = {}
 ): Promise<FacebookLoginResult> {
@@ -305,6 +327,8 @@ export async function setUserIDAsync(userID: string | null): Promise<void> {
 }
 
 /**
+ * Gets the user ID.
+ *
  * @return A promise fulfilled with the user id or null if not set.
  */
 export async function getUserIDAsync(): Promise<string | null> {
@@ -315,6 +339,8 @@ export async function getUserIDAsync(): Promise<string | null> {
 }
 
 /**
+ * Get an anonymous ID from Facebook.
+ *
  * @return A promise fulfilled with an anonymous id or null if not set.
  */
 export async function getAnonymousIDAsync(): Promise<string | null> {
@@ -325,6 +351,8 @@ export async function getAnonymousIDAsync(): Promise<string | null> {
 }
 
 /**
+ * Get the advertiser ID from Facebook.
+ *
  * @return A promise fulfilled with the advertiser id or null if not set.
  */
 export async function getAdvertiserIDAsync(): Promise<string | null> {
@@ -335,7 +363,8 @@ export async function getAdvertiserIDAsync(): Promise<string | null> {
 }
 
 /**
- * **Android only.**
+ * **Android only.** Gets the attribution ID from Facebook.
+ *
  * @return A promise fulfilled with the attribution id or null if not set.
  */
 export async function getAttributionIDAsync(): Promise<string | null> {