+++ packages/expo-sms/android/src/test/java/expo/modules/sms/MockActivityProvider.kt
@@ -1,0 +1,6 @@
+package expo.modules.sms
+
+import org.unimodules.core.interfaces.ActivityProvider
+import org.unimodules.core.interfaces.InternalModule
+
+internal interface MockActivityProvider : ActivityProvider, InternalModule