+++ packages/expo/src/logs/LogSerialization.ts
@@ -194,7 +194,11 @@ function _captureConsoleStackTrace(): Error {
 }
 
 function _getProjectRoot(): string | null {
-  return Constants.manifest?.developer?.projectRoot ?? null;
+  return (
+    Constants.manifest?.developer?.projectRoot ??
+    Constants.manifest2?.extra?.expoGo?.developer?.projectRoot ??
+    null
+  );
 }
 
 export default {